Static spots data for Preview.
"It just works" but not in a nice or dynamic way.
Now a character has to be changd in the app.js/ routes.get methods so there is no conflict and something gets displayed. Both DB & JSON GETs have route "/", now the DB's has an addidional character so the route "doesn't exist", thus there's no major conflict.