Json lg ok ; prev view kindof fixed.
Login is possible if correct person data exists in the staticPeople.json file.

(*)However for now, regardless of who logs in the displayed current-user reservated spots (green on top) willl always be the same.

Still getting unreachable json error in the browser even though the dasta is read and ( * mostly) shown as expected.